SSR the homepage cards instead of CSR fetch waterfall
The homepage previously rendered an empty shell, then fired four
sequential client requests (status, preferences, feed, ceremony banner)
before the user saw any populated content — adding ~3 RTTs to LCP for
every authenticated load.

- Extract /api/feed GET body into a reusable getFeedPagePayload helper
  (src/server/feed/get-feed-page.ts) so server components can build the
  same payload without an HTTP hop. Route becomes a thin wrapper.
- Rewrite src/app/page.tsx as a streaming server component: each card
  is an async section wrapped in Suspense, fetching its data in
  parallel server-side and passing it as initial props.
- Add initialStatus/initialPreferences to TodaysFiveCard and
  initialPage/initialBanner to FeedList. Both components fall back to
  the existing client-side fetch path when the props are absent, so
  /feed and unauthenticated /  keep working unchanged.
- Filter switches and infinite scroll still hit /api/feed.
